Move measure to Scheduler utils
