+++ qiskit/pulse/__init__.py
@@ -91,4 +91,3 @@ from .exceptions import PulseError
 from .instruction_schedule_map import InstructionScheduleMap
 from .interfaces import ScheduleComponent
 from .schedule import Schedule
-from .utils import measure, measure_all
